FALLOUT2: Add enum to int casts

+++ engines/fallout2/game_dialog.cpp
@@ -947,9 +947,9 @@ int _gdialogExitFromScript() {
 
 	GameMode::exitGameMode(GameMode::kDialog);
 
-	GameMode::enterGameMode(GameMode::kSpecial);
+	GameMode::enterGameMode((int)GameMode::kSpecial);
 	_gdDestroyHeadWindow();
-	GameMode::exitGameMode(GameMode::kSpecial);
+	GameMode::exitGameMode((int)GameMode::kSpecial);
 
 	// CE: Fix Barter button.
 	gameDialogRedButtonsExit();
@@ -1179,13 +1179,13 @@ void _gdialogUpdatePartyStatus() {
 	// NOTE: Uninline.
 	gdHide();
 
-	GameMode::enterGameMode(GameMode::kSpecial);
+	GameMode::enterGameMode((int)GameMode::kSpecial);
 
 	_gdialog_window_destroy();
 
 	gGameDialogSpeakerIsPartyMember = isPartyMember;
 
-	GameMode::exitGameMode(GameMode::kSpecial);
+	GameMode::exitGameMode((int)GameMode::kSpecial);
 
 	_gdialog_window_create();
 
@@ -1855,7 +1855,7 @@ int _gdProcess() {
 			if (_dialogue_switch_mode == 3) {
 				_dialogue_state = 4;
 
-				GameMode::exitGameMode(GameMode::kSpecial);
+				GameMode::exitGameMode((int)GameMode::kSpecial);
 
 				inventoryOpenTrade(gGameDialogWindow, gGameDialogSpeaker, _peon_table_obj, _barterer_table_obj, gGameDialogBarterModifier);
 				_gdialog_barter_cleanup_tables();
@@ -2740,7 +2740,7 @@ void gameDialogTicker() {
 		_loop_cnt = -1;
 		_dialogue_switch_mode = 3;
 
-		GameMode::enterGameMode(GameMode::kSpecial);
+		GameMode::enterGameMode((int)GameMode::kSpecial);
 
 		_gdialog_window_destroy();
 		_gdialog_barter_create_win();
